Isolate Windows desktop process launch
