+++ lib/system/guest_agent/exec_desktop_windows.go
@@ -5,6 +5,8 @@ package main
 import (
 	"context"
 	"fmt"
+	"io"
+	"log"
 	"os"
 	"sort"
 	"strings"
@@ -12,8 +14,107 @@ import (
 	"unicode/utf16"
 	"unsafe"
 
+	pb "github.com/kernel/hypeman/lib/guest"
 	"golang.org/x/sys/windows"
 )
+
+func (s *guestServer) executeNoTTY(ctx context.Context, stream pb.GuestService_ExecServer, start *pb.ExecStart) error {
+	if len(start.Command) == 0 {
+		return fmt.Errorf("empty command")
+	}
+	if start.Session != pb.ExecSession_EXEC_SESSION_DESKTOP {
+		return s.executeSystemNoTTY(ctx, stream, start)
+	}
+
+	stdinRead, stdinWrite, err := os.Pipe()
+	if err != nil {
+		return fmt.Errorf("create desktop command stdin: %w", err)
+	}
+	stdoutFile, err := os.CreateTemp("", "hypeman-exec-stdout-*")
+	if err != nil {
+		stdinRead.Close()
+		stdinWrite.Close()
+		return fmt.Errorf("create stdout capture: %w", err)
+	}
+	defer func() {
+		stdoutFile.Close()
+		os.Remove(stdoutFile.Name())
+	}()
+	stderrFile, err := os.CreateTemp("", "hypeman-exec-stderr-*")
+	if err != nil {
+		stdinRead.Close()
+		stdinWrite.Close()
+		return fmt.Errorf("create stderr capture: %w", err)
+	}
+	defer func() {
+		stderrFile.Close()
+		os.Remove(stderrFile.Name())
+	}()
+
+	process, cleanup, err := startDesktopProcess(
+		ctx,
+		start.Command,
+		s.buildEnv(start.Env, false),
+		start.Cwd,
+		stdinRead,
+		stdoutFile,
+		stderrFile,
+		start.TimeoutSeconds,
+	)
+	stdinRead.Close()
+	if err != nil {
+		stdinWrite.Close()
+		return err
+	}
+	defer cleanup()
+	go func() {
+		defer stdinWrite.Close()
+		for {
+			req, err := stream.Recv()
+			if err != nil {
+				return
+			}
+			if data := req.GetStdin(); data != nil {
+				_, _ = stdinWrite.Write(data)
+			}
+		}
+	}()
+
+	processState, waitErr := process.Wait()
+	if _, err := stdoutFile.Seek(0, io.SeekStart); err != nil {
+		return fmt.Errorf("rewind stdout capture: %w", err)
+	}
+	stdout, err := io.ReadAll(stdoutFile)
+	if err != nil {
+		return fmt.Errorf("read stdout capture: %w", err)
+	}
+	if _, err := stderrFile.Seek(0, io.SeekStart); err != nil {
+		return fmt.Errorf("rewind stderr capture: %w", err)
+	}
+	stderr, err := io.ReadAll(stderrFile)
+	if err != nil {
+		return fmt.Errorf("read stderr capture: %w", err)
+	}
+
+	const chunkSize = 32 * 1024
+	for i := 0; i < len(stdout); i += chunkSize {
+		end := min(i+chunkSize, len(stdout))
+		_ = stream.Send(&pb.ExecResponse{Response: &pb.ExecResponse_Stdout{Stdout: stdout[i:end]}})
+	}
+	for i := 0; i < len(stderr); i += chunkSize {
+		end := min(i+chunkSize, len(stderr))
+		_ = stream.Send(&pb.ExecResponse{Response: &pb.ExecResponse_Stderr{Stderr: stderr[i:end]}})
+	}
+
+	exitCode := int32(0)
+	if processState != nil {
+		exitCode = int32(processState.ExitCode())
+	} else if waitErr != nil {
+		exitCode = 124
+	}
+	log.Printf("[guest-agent] desktop command finished with exit code: %d", exitCode)
+	return stream.Send(&pb.ExecResponse{Response: &pb.ExecResponse_ExitCode{ExitCode: exitCode}})
+}
 
 func startDesktopProcess(
 	ctx context.Context,
